Make AbilityEvents return a result
There will be a lot of times when a player
tries to use an ability but can't. For example,
it could be on cooldown.

Given this, we need to tell the caller if the
event they tried to fire succeeded or failed.

This could allow us to add behavior like voice
over lines like "I can't do that right now" when
the ability can't be used.