Implement CooldownState
This was a tough one.

You want the implementations of IPlayable to
be responsible for knowing when their abilities
are on/off cooldown, so I had to do a pub-sub
approach here. (Or you at least want them to
centralize that knowledge and delegate to another
component).

The tough part is the notifications are basically
ONLY relavant to one ability (since not everything
shares a cooldown) so we have to have different
overloads for the different abilities.

Yay! double dispatch again xD that lets things be
pretty tame from the CooldownState side since its
literally just supposed to be like "yes this is on
CD" or "No it is not"