Adds Playable Interface and Ability StateMachine (#15)
* Adds IPlayable interface for playable creatures

* Support BasicAttack events

* Support ChargedAttack events

* Use normal comments instead of method comments

* Support CharacterAbility events

* Rename ChargedAttack to ChargedAbility

It's not necessarily true that every
characters RMB needs to be an ATTACK

For example, a winger might have a
chargeable heal.

* Support CoreInfusionAbility events

* Support MobilityAbility events

* Add a prop for CoreInfusion

* Make AbilityEvents return a result

There will be a lot of times when a player
tries to use an ability but can't. For example,
it could be on cooldown.

Given this, we need to tell the caller if the
event they tried to fire succeeded or failed.

This could allow us to add behavior like voice
over lines like "I can't do that right now" when
the ability can't be used.

* Add AbilityStates

* Make it possible for states to delegate back to events

* Implement ReadyState

* Implement CooldownState

This was a tough one.

You want the implementations of IPlayable to
be responsible for knowing when their abilities
are on/off cooldown, so I had to do a pub-sub
approach here. (Or you at least want them to
centralize that knowledge and delegate to another
component).

The tough part is the notifications are basically
ONLY relavant to one ability (since not everything
shares a cooldown) so we have to have different
overloads for the different abilities.

Yay! double dispatch again xD that lets things be
pretty tame from the CooldownState side since its
literally just supposed to be like "yes this is on
CD" or "No it is not"

* Implement DisabledState

* Add newlines to EoFs

* Add method docs to IStateNotifier

---------

Co-authored-by: iNDirection <Noah.Denier@gmail.com>