Remove double-dispatch from StateNotifier pub-sub logic
It would make sense to use double dispatch here if
we wanted different implementations of player chars
to be able to track cooldowns in their own way on
an ability by ability basis ... which sounded like a
bad kind of flexibility once I started implementing.

The thing I wanted was for each ability to have its
own cooldown value, and there was a simpler way to
accomplish that so I removed the per ability
Subscribe/Unsubscribe methods.